pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["engine"]

[project]
name = "timelapse-engine"
version = "0.1.0"
description = "Hexagonal Architecture Timelapse Engine"
authors = [
    {name = "Raka Arwaky", email = "raka@example.com"},
]
dependencies = [
    "networkx",
    "Pillow",
    "numpy",
    "psd-tools",
    "moviepy",
    "fastapi",
    "uvicorn",
    "pydantic",
    "python-multipart",
    "bezier"
]
requires-python = ">=3.10"

[tool.ruff]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E", "F", "I", "N", "PL", "RUF"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 100

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
mypy_path = "engine"
explicit_package_bases = true

[[tool.mypy.overrides]]
module = [
    "networkx",
    "cv2",
    "PIL.*",
    "psd_tools.*",
    "moviepy.*",
    "bezier"
]
ignore_missing_imports = true
